fix benchmarking issue by using `lib.rs`
