+++ src/main.rs
@@ -1,59 +1,9 @@
-use ahash::AHashMap;
-use clap::Parser;
 use std::fs;
+use clap::Parser;
 
-mod output;
-
-mod stats;
-
-mod generation;
-
-#[derive(PartialEq)]
-pub struct Key {
-    hand: u8,
-    finger: Finger,
-    row: u8,
-    lateral: bool,
-}
-
-#[derive(Eq, Hash, PartialEq, PartialOrd)]
-enum Finger {
-    Thumb,
-    Index,
-    Middle,
-    Ring,
-    Pinky,
-}
-
-#[derive(Default, Debug, Clone)]
-pub struct Stats {
-    score: i64,
-    fspeed: i64,
-    sfb: i64,
-    sfr: i64,
-    sfs: i64,
-    lsb: i64,
-    lss: i64,
-    fsb: i64,
-    fss: i64,
-    inroll: i64,
-    outroll: i64,
-    alt: i64,
-    inthreeroll: i64,
-    outthreeroll: i64,
-    weak_red: i64,
-    red: i64,
-    heatmap: i64,
-    thumb_stat: u32,
-    pub bigrams: u32,
-    pub skipgrams: u32,
-    pub trigrams: u32,
-    pub ngram_table: AHashMap<[char; 3], u32>,
-    pub bad_bigrams: Vec<String>,
-}
+use chogalyzer::*;
 
 #[derive(Parser, Debug)]
-
 struct Args {
     #[arg(short, long, default_value = "whirl.txt")]
     layout: String,
@@ -73,10 +23,6 @@ struct Args {
     #[arg(long, default_value_t = 0.99)]
     cooling: f64,
 }
-
-const INCLUDE_THUMB_ALT: bool = true;
-const INCLUDE_THUMB_ROLL: bool = true;
-const INCLUDE_SPACE: bool = true;
 
 fn main() {
     let args = Args::parse();
